Refactor for tests: start work on Emitter.
